Added dev-libs/icu to depends dev-db/postgresql-base when the 1c USE flag is enabled. Minor clean ups.

+++ dev-db/postgresql-base/postgresql-base-9.0.3-r2.ebuild
@@ -1,32 +1,34 @@
-# Copyright 1999-2011 Gentoo Foundation
+# Copyright 1999-2012 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-db/postgresql-base/postgresql-base-9.0.3.ebuild,v 1.3 2011/02/05 11:41:38 fauli Exp $
+# $Header: $
 
-EAPI="2"
+EAPI="3"
 
 WANT_AUTOMAKE="none"
 
-inherit eutils multilib versionator autotools prefix
+inherit autotools eutils multilib prefix versionator
 
-KEYWORDS="~alpha ~amd64 ~arm hppa ~ia64 ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc x86 ~sparc-fbsd ~x86-fbsd"
+SLOT="$(get_version_component_range 1-2)"
+
+KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86 ~sparc-fbsd ~x86-fbsd"
 
 DESCRIPTION="PostgreSQL libraries and clients"
 HOMEPAGE="http://www.postgresql.org/"
-
-MY_PV=${PV/_/}
-SRC_URI="mirror://postgresql/source/v${MY_PV}/postgresql-${MY_PV}.tar.bz2"
-S=${WORKDIR}/postgresql-${MY_PV}
-
+SRC_URI="mirror://postgresql/source/v${PV}/postgresql-${PV}.tar.bz2
+		 http://dev.gentoo.org/~titanofold/postgresql-patches-${SLOT}.tbz2"
 LICENSE="POSTGRESQL"
-SLOT="$(get_version_component_range 1-2)"
+
+S=${WORKDIR}/postgresql-${PV}
+
+# No tests to be done for clients and libraries
+RESTRICT="test"
+
 LINGUAS="af cs de es fa fr hr hu it ko nb pl pt_BR ro ru sk sl sv tr zh_CN zh_TW"
-IUSE="doc kerberos ldap nls pam +pg_legacytimestamp readline ssl threads zlib"
+IUSE="1c doc kerberos ldap nls pam pg_legacytimestamp readline ssl threads zlib"
 
 for lingua in ${LINGUAS}; do
 	IUSE+=" linguas_${lingua}"
 done
-
-RESTRICT="test"
 
 wanted_languages() {
 	local enable_langs
@@ -57,16 +59,20 @@ DEPEND="${RDEPEND}
 PDEPEND="doc? ( ~dev-db/postgresql-docs-${PV} )"
 
 src_prepare() {
-	epatch "${FILESDIR}/postgresql-9.0-common.3.patch" \
-		"${FILESDIR}/postgresql-${SLOT}-base.3.patch" \
-                "${FILESDIR}/1c_postgresql-9.0-logging.patch" \
-                "${FILESDIR}/1c_postgresql-perl-rpath.patch" \
-                "${FILESDIR}/1c_postgresql-prefer-ncurses.patch" \
-                "${FILESDIR}/1c_FULL_90-0.19.patch" \
-                "${FILESDIR}/1c_postgresql-9.0-configs.patch" \
-                "${FILESDIR}/1c_postgresql-9.0-applock.patch" || die "epatch failed"
 
+	epatch "${WORKDIR}/autoconf.patch" \
+		"${WORKDIR}/base.patch" || die "epatch failed"
 
+	if use 1c ; then
+	    epatch "${FILESDIR}/1c_postgresql-9.0-logging.patch" \
+		"${FILESDIR}/1c_postgresql-perl-rpath.patch" \
+		"${FILESDIR}/1c_postgresql-prefer-ncurses.patch" \
+		"${FILESDIR}/1c_FULL_90-0.19.patch" \
+		"${FILESDIR}/1c_postgresql-9.0-r1-configs.patch" \
+		"${FILESDIR}/1c_postgresql-9.0-applock.patch" || die "1c patch set failed"
+	fi
+
+	eprefixify src/include/pg_config_manual.h
 
 	# to avoid collision - it only should be installed by server
 	rm "${S}/src/backend/nls.mk"
@@ -80,12 +86,12 @@ src_prepare() {
 src_configure() {
 	export LDFLAGS_SL="${LDFLAGS}"
 	econf \
-		--prefix=/usr/$(get_libdir)/postgresql-${SLOT} \
-		--datadir=/usr/share/postgresql-${SLOT} \
-		--docdir=/usr/share/doc/postgresql-${SLOT} \
-		--sysconfdir=/etc/postgresql-${SLOT} \
-		--includedir=/usr/include/postgresql-${SLOT} \
-		--mandir=/usr/share/postgresql-${SLOT}/man \
+		--prefix=${EROOT%/}/usr/$(get_libdir)/postgresql-${SLOT} \
+		--datadir=${EROOT%/}/usr/share/postgresql-${SLOT} \
+		--docdir=${EROOT%/}/usr/share/doc/postgresql-${SLOT} \
+		--sysconfdir=${EROOT%/}/etc/postgresql-${SLOT} \
+		--includedir=${EROOT%/}/usr/include/postgresql-${SLOT} \
+		--mandir=${EROOT%/}/usr/share/postgresql-${SLOT}/man \
 		--enable-depend \
 		--without-tcl \
 		--without-perl \
@@ -115,9 +121,9 @@ src_install() {
 	doins "${S}"/src/include/postmaster/*.h || die
 
 	dodir /usr/share/postgresql-${SLOT}/man/man1/ || die
-	cp  "${S}"/doc/src/sgml/man1/* "${D}"/usr/share/postgresql-${SLOT}/man/man1/ || die
+	cp  "${S}"/doc/src/sgml/man1/* "${ED}"/usr/share/postgresql-${SLOT}/man/man1/ || die
 
-	rm "${D}/usr/share/postgresql-${SLOT}/man/man1"/{initdb,ipcclean,pg_controldata,pg_ctl,pg_resetxlog,pg_restore,postgres,postmaster}.1
+	rm "${ED}/usr/share/postgresql-${SLOT}/man/man1"/{initdb,pg_controldata,pg_ctl,pg_resetxlog,pg_restore,postgres,postmaster}.1 || die
 	dodoc README HISTORY doc/{README.*,TODO,bug.template} || die
 
 	cd "${S}/contrib"
@@ -126,24 +132,24 @@ src_install() {
 
 	dodir /etc/eselect/postgresql/slots/${SLOT} || die
 
-	IDIR="/usr/include/postgresql-${SLOT}"
-	cat > "${D}/etc/eselect/postgresql/slots/${SLOT}/base" <<-__EOF__
+	IDIR="${EROOT%/}/usr/include/postgresql-${SLOT}"
+	cat > "${ED}/etc/eselect/postgresql/slots/${SLOT}/base" <<-__EOF__
 postgres_ebuilds="\${postgres_ebuilds} ${PF}"
-postgres_prefix=/usr/$(get_libdir)/postgresql-${SLOT}
-postgres_datadir=/usr/share/postgresql-${SLOT}
-postgres_bindir=/usr/$(get_libdir)/postgresql-${SLOT}/bin
+postgres_prefix=${EROOT%/}/usr/$(get_libdir)/postgresql-${SLOT}
+postgres_datadir=${EROOT%/}/usr/share/postgresql-${SLOT}
+postgres_bindir=${EROOT%/}/usr/$(get_libdir)/postgresql-${SLOT}/bin
 postgres_symlinks=(
-	${IDIR} /usr/include/postgresql
-	${IDIR}/libpq-fe.h /usr/include/libpq-fe.h
-	${IDIR}/pg_config_manual.h /usr/include/pg_config_manual.h
-	${IDIR}/libpq /usr/include/libpq
-	${IDIR}/postgres_ext.h /usr/include/postgres_ext.h
+	${IDIR} ${EROOT%/}/usr/include/postgresql
+	${IDIR}/libpq-fe.h ${EROOT%/}/usr/include/libpq-fe.h
+	${IDIR}/pg_config_manual.h ${EROOT%/}/usr/include/pg_config_manual.h
+	${IDIR}/libpq ${EROOT%/}/usr/include/libpq
+	${IDIR}/postgres_ext.h ${EROOT%/}/usr/include/postgres_ext.h
 )
 __EOF__
 
 	cat >"${T}/50postgresql-94-${SLOT}" <<-__EOF__
-LDPATH=/usr/$(get_libdir)/postgresql-${SLOT}/$(get_libdir)
-MANPATH=/usr/share/postgresql-${SLOT}/man
+LDPATH=${EROOT%/}/usr/$(get_libdir)/postgresql-${SLOT}/$(get_libdir)
+MANPATH=${EROOT%/}/usr/share/postgresql-${SLOT}/man
 __EOF__
 	doenvd "${T}/50postgresql-94-${SLOT}" || die
 
@@ -154,8 +160,7 @@ pkg_postinst() {
 	eselect postgresql update
 	[[ "$(eselect postgresql show)" = "(none)" ]] && eselect postgresql set ${SLOT}
 	elog "If you need a global psqlrc-file, you can place it in:"
-	elog "    '${ROOT}/etc/postgresql-${SLOT}/'"
-	elog
+	elog "    ${EROOT%/}/etc/postgresql-${SLOT}/"
 }
 
 pkg_postrm() {
